refactor: new folder structure for tests
