+++ src/__tests__/integration/routes/login/post.spec.ts
@@ -1,9 +1,9 @@
 import request from 'supertest';
 import bcrypt from 'bcrypt';
-import app from '../../../config/app.config';
-import prisma from '../../../config/prisma.config';
-import { users as usersMock } from '../../../mocks/users.mock';
-import { IResponse } from '../../../types';
+import app from '../../../../configs/app.config';
+import prisma from '../../../../configs/prisma.config';
+import { users as usersMock } from '../../../../mocks/users.mock';
+import { IResponse } from '../../../../types';
 
 afterEach(() => {
   jest.resetAllMocks();
