+++ src/__tests__/integration/routes/posts/get.spec.ts
@@ -1,7 +1,7 @@
-import app from '../../../config/app.config';
+import app from '../../../../configs/app.config';
 import request from 'supertest';
-import prisma from '../../../config/prisma.config';
-import { IResponse } from '../../../types';
+import prisma from '../../../../configs/prisma.config';
+import { IResponse } from '../../../../types';
 import { Post } from '@prisma/client';
 
 describe('GET /posts', () => {
