+++ src/__tests__/integration/routes/users/put.spec.ts
@@ -1,9 +1,9 @@
 import request from 'supertest';
 import { Prisma } from '@prisma/client';
-import app from '../../../config/app.config';
-import prisma from '../../../config/prisma.config';
-import { IResponse } from '../../../types';
-import { users as usersMock } from '../../../mocks/users.mock';
+import app from '../../../../configs/app.config';
+import prisma from '../../../../configs/prisma.config';
+import { IResponse } from '../../../../types';
+import { users as usersMock } from '../../../../mocks/users.mock';
 import jwt from 'jsonwebtoken';
 
 afterEach(() => {
