+++ src/controllers/login.controllers.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 import jwt from 'jsonwebtoken';
 import bcrypt from 'bcrypt';
-import prisma from '../config/prisma.config';
+import prisma from '../configs/prisma.config';
 import { Controller, ILoginInfo } from '../types';
 
 export const login: Controller = async (req: Request, res: Response) => {
