+++ src/index.ts
@@ -1,4 +1,4 @@
-import app from './config/app.config';
+import app from './configs/app.config';
 
 const port = process.env.SERVER_PORT || 3030;
